TextUi.java: Apply SRP and extract it to Formatter.java
